+++ CHANGELOG.md
@@ -4,7 +4,11 @@
 
 ### TL;DR
 
-`pyannote.audio` does [speech separation](https://arxiv.org/abs/2403.02288): multi-speaker audio in, one audio channel per speaker out
+`pyannote.audio` does [speech separation](https://hf.co/pyannote/speech-separation-ami-1.0): multi-speaker audio in, one audio channel per speaker out!
+
+```bash
+pip install pyannote.audio[separation]==3.3.0
+```
 
 ### New features